Added ZMQ Socket Monitoring & Poller Timeouts
This update adds ZeroMQ socket monitoring to the class, whereby PUSH and PULL operations are not attempted if a successful EVENT_HANDSHAKE has not yet been triggered on either the PUSH or PULL socket.

The user is now informed of the current state of either of the two sockets. Previously users would receive no output, obviously leading to confusion as to what was going on behind the scenes.